Fix JSON formatting, remove _parseAndPrint, fix formatting
The JSON ouput from the script was broken for all servers except
the Node server. The JSON was being escaped prematurely. The Node
server now uses a stringified object instead of manually constructing
a JSON object removing the need to escape from the script file.

The Node server didn't use semicolons and had inconsistent formatting,
that is now fixed.

The Node server can now accept a single CLI argument which is the port
number to bind the server.